增加UI（状态页，设置页）

+++ src/dvb/mytype.h
@@ -3,12 +3,12 @@
 
 typedef char int8;
 typedef short int16;
-typedef long int32;
+typedef int int32;
 typedef long long int64;
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned long uint32;
+typedef unsigned int uint32;
 typedef unsigned long long uint64;
 
 #ifndef NULL
